change decimals to floats in db

+++ db/migrate/20170310010439_create_heroes.rb
@@ -6,11 +6,11 @@ class CreateHeroes < ActiveRecord::Migration[5.0]
       t.integer :base_str
       t.integer :base_agi
       t.integer :base_int
-      t.decimal :add_str
-      t.decimal :add_agi
-      t.decimal :add_int
+      t.float :add_str
+      t.float :add_agi
+      t.float :add_int
       t.integer :move_speed
-      t.decimal :turn_rate
+      t.float :turn_rate
       t.string :attack_type
       t.jsonb :talents
       t.integer :attack_range
